pkg/controller: Refactor scheduler - move scheduler metrics and hooks to dedicate files - remove weight params from shouldSkipAnalysis
